requirements.txt
streamlit
streamlit-elements==0.1.
pandas
plotly.express
protobuf==3.13
